add actions

+++ src/features/match/details/teams/MatchPlayer.tsx
@@ -5,6 +5,8 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { ErrorMessages } from "../../../../util/constants";
 import { userHasAccess } from "../../../../util/sec";
 import MatchService from "../../../../services/MatchService";
+import { useState } from "react";
+import PlayerActionsModal from "./PlayerActionsModal";
 
 interface MatchPlayerProps {
 	match: IMatch;
@@ -17,12 +19,14 @@ const getTeam = (player: IMatchPlayer) => {
 };
 
 function MatchPlayer(props: MatchPlayerProps) {
+	const [showPlayerActions, setShowPlayerActions] = useState<boolean>(false);
 	const { match, player, onActionSuccess } = props;
 
 	const removePlayer = () => {
 		if (userHasAccess()) {
 			const onSuccess = (response: any) => {
 				if (response.status === 200) {
+					setShowPlayerActions(false);
 					onActionSuccess();
 				}
 			};
@@ -38,6 +42,7 @@ function MatchPlayer(props: MatchPlayerProps) {
 	const addGoal = () => {
 		const onSuccess = (response: any) => {
 			if (response.status === 200) {
+				setShowPlayerActions(false);
 				onActionSuccess();
 			}
 		};
@@ -50,56 +55,47 @@ function MatchPlayer(props: MatchPlayerProps) {
 	};
 
 	const removeGoal = () => {
-		const onSuccess = (response: any) => {
-			if (response.status === 200) {
-				onActionSuccess();
-			}
-		};
+		if (userHasAccess()) {
+			const onSuccess = (response: any) => {
+				if (response.status === 200) {
+					setShowPlayerActions(false);
+					onActionSuccess();
+				}
+			};
 
-		const onError = (response: any) => {
-			alert(ErrorMessages.GENERIC);
-		};
+			const onError = (response: any) => {
+				alert(ErrorMessages.GENERIC);
+			};
 
-		MatchService.removeGoal(match.match_id, { player_name: player.player_name }, onSuccess, onError);
-	};
-
-	const handlePlayerAction = () => {
-		const action = prompt("O que deseja fazer? 1 (+ 1 gol), 2(- 1 gol), 3 (remover jogador)");
-
-		const actionNumber = Number(action);
-
-		if (actionNumber === null || isNaN(actionNumber) || actionNumber < 1 || actionNumber > 3) {
-			alert("Opção inválida");
-		} else {
-			switch (actionNumber) {
-				case 1:
-					addGoal();
-					break;
-				case 2:
-					removeGoal();
-					break;
-				default:
-					removePlayer();
-					break;
-			}
+			MatchService.removeGoal(match.match_id, { player_name: player.player_name }, onSuccess, onError);
 		}
 	};
 
 	return (
-		<button
-			disabled={match.is_ended}
-			onClick={() => handlePlayerAction()}
-			className={`team-player-container team-player-button teams-${getTeam(player)}-team disabled-${match.is_ended}`}
-		>
-			<div className="team-player-name">{player.player_name}</div>
-			<div className="team-player-goals">
-				{player.goals_scored === 0 && !match.is_ended ? (
-					<FontAwesomeIcon icon={faPlus} />
-				) : (
-					[...Array(player.goals_scored)].map((e, i) => <FontAwesomeIcon key={i} icon={faFutbol} />)
-				)}
-			</div>
-		</button>
+		<div>
+			<button
+				disabled={match.is_ended}
+				onClick={() => setShowPlayerActions(true)}
+				className={`team-player-container team-player-button teams-${getTeam(player)}-team disabled-${match.is_ended}`}
+			>
+				<div className="team-player-name">{player.player_name}</div>
+				<div className="team-player-goals">
+					{player.goals_scored === 0 && !match.is_ended ? (
+						<FontAwesomeIcon icon={faPlus} />
+					) : (
+						[...Array(player.goals_scored)].map((e, i) => <FontAwesomeIcon key={i} icon={faFutbol} />)
+					)}
+				</div>
+			</button>
+			{showPlayerActions && (
+				<PlayerActionsModal
+					onClose={() => setShowPlayerActions(false)}
+					onAddGoal={() => addGoal()}
+					onRemoveGoal={() => removeGoal()}
+					onRemovePlayer={() => removePlayer()}
+				></PlayerActionsModal>
+			)}
+		</div>
 	);
 }
